+++ packages/app/src/app/pages/Sandbox/Editor/Workspace/items/Live/LiveInfo/Preferences/index.ts
@@ -1,0 +1,1 @@
+export { Preferences } from './Preferences';
