+++ packages/app/src/app/pages/Sandbox/Editor/Workspace/items/Live/LiveInfo/types.ts
@@ -1,0 +1,5 @@
+export type Theme = {
+  theme: {
+    light: boolean;
+  };
+};
